(bug 578) Fix installer on Ubuntu 18.04
We originally ran into this at the hackfest, where the installer would
fail on Ubuntu 18.04, the latest long-term supported release.

The behaviour of the `sg` command has failed (an apparent upstream
change) so that it can't be reliably used to check the error code of the
command it runs.

Considering the original use of `sg`, to ensure that the user had access
to lxd in order to bootstrap, reduces how we use `sg`, only calling it
if we need to bootstrap a new juju controller.

The other commands -- `juju show-controller` and `juju controllers` --
do not need access to the lxd group in order to function.

I've tested this patch on Ubuntu 16.04 and 18.04 and verified that it
works as expected on both (where lxd now pre-installed) and will also
work with older versions of 16.04 that may not have it installed.

Signed-off-by: Adam Israel <adam.israel@canonical.com>
Change-Id: I0e50ef3ac3cf127b63f5f56d8f90fea44f5be0f2