validate_descriptor.py: new option for validate charms
Default behaviour is NOT to validate whether the charm is in the folder.
The rationale is that this tool is intended to validate only the descriptor,
not the package. Validating the package requires more aspects to be
taken into account, not only the charm
Moreover, in a CICD for a VNF vendor, charms can be created and tested
independently from the descriptors and only be merged in the same folder
in the last step.

Change-Id: I33aa986493931df463fa8f9d2c6cda2e4be45cf9
Signed-off-by: garciadeblas <gerardo.garciadeblas@telefonica.com>