Fix pip version dependency
There are numerous package differences between the Docker image built by
Jenkins and the Dockerfile.local in the LCM repository. One of these
version mismatches is causing native charms in 6.0.4rc1 to fail.

In particular, the version of cffi needs to be upgraded in order for the
paramiko library to successfully ssh to the VM being used to run the
native charm.

This patch installs the current stable version, 1.13.2, of cffi, rather
than the previously installed version of 1.5.2

Change-Id: If200314a16b652562bb543c0415de4e71c0bcb09
Signed-off-by: Adam Israel <adam.israel@canonical.com>