Make Charmed Installer Idempotent
By default, the charmed installer will create a controller called
'controller', unless an external VCA is specified using the --vca
option.  Upon uninstall, it is not known or safe to remove this
Juju controller as there might be other applications using it.

This change shows that message on uninstall, giving the user the
opportunity to remove it with a command.

Also, as part of the installation, this change checks for the
controller and if it exists, presents the user with options for
how to resolve the situation.

Fixes bug 1126

Change-Id: Id13cd822ed7eef60250544108b74bbba7998d9b7
Signed-off-by: beierlm <mark.beierl@canonical.com>